fix(feature-provider): add missing export module
