--- packages/bezier-react/src/components/FeatureProvider/Feature/Feature.types.ts
@@ -1,8 +1,0 @@
-export enum FeatureType {
-  SmoothCorners = 'smooth-corners',
-}
-
-export interface Feature {
-  readonly name: FeatureType
-  activate: () => Promise<boolean>
-}
