--- packages/bezier-react/src/components/FeatureProvider/Feature/index.ts
@@ -1,5 +1,0 @@
-export {
-  type Feature,
-  FeatureType,
-} from './Feature.types'
-export { SmoothCornersFeature } from './SmoothCornersFeature'
