+++ packages/bezier-react/src/components/FeatureProvider/SmoothCornersFeature/smoothCornersScript.ts
